Fix X11 GUI stuff
	modified:   ../cvt/gl/GFXEngineGL.cpp
	modified:   ../cvt/gl/GFXEngineGLFBO.h
	modified:   ../cvt/gui/event/KeyEvent.h
	modified:   ../cvt/gui/internal/X11/ApplicationX11.cpp
	modified:   ../cvt/gui/internal/X11/WidgetImplWinGLX11.cpp

+++ cvt/gui/event/KeyEvent.h
@@ -62,7 +62,7 @@ namespace cvt {
 			out << " Meta";
 		out << std::endl;
 
-		out << "\tCode: " << std::hex << "0x" << k.key() << std::endl;
+		out << "\tCode: " << std::hex << "0x" << ( unsigned int ) k.key() << std::endl;
 		out << "\tText: " << "'" <<k.text() << "'";
         return out;
     }
